requirements.txt
aiohttp
aiofiles
azure-functions
bonnette
fastapi
azure-storage-blob